Update __init__.py
From April 1st Glitre is reducing the grip price for both winter and summer half year.
They have also changed the night reduction to be -0.12 NOK all year. Changed the code to reflect this.
Price for winter half year is not changed yet. That can be done after April 1st.