Glitre is again changing grid price
Glitre is again changing grid price. But not until May 1st.
As the price list says just now there is no difference on summer and winter prices this year. I am not sure if this is only a temporary adjustment, so therefore the logic isn't removed completely for winter/summer price.